+++ src/tp1/server/resource/SpreadsheetsResource.java
@@ -130,7 +130,7 @@ public class SpreadsheetsResource implements RestSpreadsheets {
 	}
 
 	@Override
-	public Spreadsheet getSpreadsheet(String sheetId, String userId, String password, int version) {
+	public Spreadsheet getSpreadsheet(String sheetId, String userId, String password, Long version) {
 		Log.info("getSpreadsheet : " + sheetId + "; userId = " + userId + "; pwd = " + password);
 
 		if (sheetId == null || userId == null)
@@ -174,7 +174,7 @@ public class SpreadsheetsResource implements RestSpreadsheets {
 	}
 
 	@Override
-	public String[][] getSpreadsheetValues(String sheetId, String userId, String password, int version) {
+	public String[][] getSpreadsheetValues(String sheetId, String userId, String password, Long version) {
 		Log.info("getSpreadsheetValues : " + sheetId + "; userId = " + userId + "; pwd = " + password);
 
 		if (sheetId == null || userId == null || password == null)
@@ -397,7 +397,7 @@ public class SpreadsheetsResource implements RestSpreadsheets {
 	}
 
 	@Override
-	public String[][] importRange(String sheetId, String userId, String range, String secret, int version) {
+	public String[][] importRange(String sheetId, String userId, String range, String secret, Long version) {
 		Log.info("importRange : " + sheetId + "; userId = " + userId + "; range = " + range);
 
 		if(!secret.equals(SpreadsheetsServer.serverSecret))
